[OPS-2598] Code redesign
